% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{equivalency_AUC_any}
\alias{equivalency_AUC_any}
\title{Sample size for testing equivalency when the accuracy is AUC
under any distribution}
\usage{
equivalency_AUC_any(theta1, theta2, r, Delta, alpha, beta, R)
}
\arguments{
\item{theta1}{AUC of test 1}

\item{theta2}{AUC of test 2}

\item{r}{the correlation between the tests because of the paired design}

\item{Delta}{the smallest difference in accuracy which is not acceptable}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size(paired design)
}
\description{
Sample size for testing equivalency when the accuracy is AUC
under any distribution
}
