% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{equivalency_AUC_binormal}
\alias{equivalency_AUC_binormal}
\title{Sample size for testing equivalency when the accuracy is AUC
under the binormal assumption}
\usage{
equivalency_AUC_binormal(a1, b1, a2, b2, rD, rN, Delta, alpha, beta, R)
}
\arguments{
\item{a1}{the binormal parameter a for test 1 under the alternative hypothesis}

\item{b1}{the binormal parameter b for test 1 under the alternative hypothesis}

\item{a2}{the binormal parameter a for test 2 under the alternative hypothesis}

\item{b2}{the binormal parameter b for test 2 under the alternative hypothesis}

\item{rD}{the correlation of the underlying bivariate binormal distribution for patients with the condition}

\item{rN}{the correlation of the underlying bivariate binormal distribution for patients without the condition}

\item{Delta}{the smallest difference in accuracy which is not acceptable}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size(paired design)
}
\description{
Sample size for testing equivalency when the accuracy is AUC
under the binormal assumption
}
