% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{equivalency_fixedFPR}
\alias{equivalency_fixedFPR}
\title{Sample size for testing equivalency when the accuracy is sensitivity at fixed FPR
We compare the z_transformed sensitivity}
\usage{
equivalency_fixedFPR(a1, b1, a2, b2, rD, rN, e, Delta, alpha, beta, R)
}
\arguments{
\item{a1}{binormal parameter for test 1 under alternative hypothesis}

\item{b1}{binormal parameter for test 1 under alternative hypothesis}

\item{a2}{binormal parameter for test 2 under alternative hypothesis}

\item{b2}{binormal parameter for test 2 under alternative hypothesis}

\item{rD}{the correlation of the underlying bivariate binormal distribution for patients with the condition}

\item{rN}{the correlation of the underlying bivariate binormal distribution for patients without the condition}

\item{e}{the fixed FPR rate}

\item{Delta}{the smallest difference in accuracy which is not acceptable}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size(paired design)
}
\description{
Sample size for testing equivalency when the accuracy is sensitivity at fixed FPR
We compare the z_transformed sensitivity
}
