% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.1.5.R
\name{fixedFPRsample}
\alias{fixedFPRsample}
\title{Sample Size for Estimating Sensitivity at Fixed FPR}
\usage{
fixedFPRsample(a0, b0, e, L, alpha, beta, R)
}
\arguments{
\item{a0}{the parameters of the assumed underlying binormal distribution}

\item{b0}{the parameters of the assumed underlying binormal distribution}

\item{e}{the fixed FPR rate}

\item{L}{half of the length of CI for the sensitivity}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n_one sample size of diseased subjects in a one-sided hypothesis testing
        N_one total sample size in a one-sided hypothesis testing
        n_two sample size of diseased subjects in a two-sided hypothesis testing
        N_two total sample size in a two-sided hypothesis testing
}
\description{
Sample Size for Estimating Sensitivity at Fixed FPR
}
