% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.1.4.R
\name{hypoarea}
\alias{hypoarea}
\title{Testing the Hypothesis that the ROC Area is Equal to a Particular Value}
\usage{
hypoarea(null, alter, alpha, beta, R)
}
\arguments{
\item{null}{area under the null hypothesis}

\item{alter}{area under the alternative hypothesis}

\item{alpha}{significance level}

\item{beta}{1-beta is power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n_one sample size of diseased subjects in a one-sided hypothesis testing
        N_one total sample size in a one-sided hypothesis testing
        n_two sample size of diseased subjects in a two-sided hypothesis testing
        N_two total sample size in a two-sided hypothesis testing
}
\description{
Testing the Hypothesis that the ROC Area is Equal to a Particular Value
}
