% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.5.1.R
\name{MRMC_nopilot_partialAUC}
\alias{MRMC_nopilot_partialAUC}
\title{calculate the power of the test with J readers and n individuals with the condition when pilot data is not available
the accuracy is partial AUC}
\usage{
MRMC_nopilot_partialAUC(
  a,
  b,
  e1,
  e2,
  delta,
  var_b,
  var_w,
  rho_1,
  rho_2,
  rho_3,
  rho_b,
  J,
  n,
  Q,
  R,
  alpha
)
}
\arguments{
\item{a}{binormal parameter a under the null hypothesis}

\item{b}{binormal parameter b under the null hypothesis}

\item{e1}{lower bound for FPR of partial area}

\item{e2}{upper bound for FPR of partial area}

\item{delta}{the difference of the mean of partial area index under the alternative hypothesis}

\item{var_b}{The estimated variability from different readers interpreting the results of the same patients using the same test}

\item{var_w}{The estimated variability from the same reader interpreting the results of the same patients using the same test on different occasions}

\item{rho_1}{The estimated correlation when the same patients are evaluated by the same reader using different tests}

\item{rho_2}{The estimated correlation when the same patients are evaluated by different readers using the same test}

\item{rho_3}{The estimated correlation when the same patients are evaluated by different readers using different tests}

\item{rho_b}{The estimated correlation when the same readers evaluate patients using different tests}

\item{J}{number of readers for each diagnostic test}

\item{n}{number of individuals with the condition
For the paired-patient study designs,n is the total number of individuals with the condition needed.
For unpaired-patient designs,n is the number of individuals with the condition needed per diagnostic test.
For the paired-patient per reader design,n individuals with the condition are needed for each of the J readers.}

\item{Q}{number of times each reader interprets the test results of each patient using the same test (often Q=1)}

\item{R}{the ratio of number of individuals without the condition and the number of individuals with the condition}

\item{alpha}{significance level}
}
\value{
the power of the test with J readers and n individuals with the condition when pilot data is not available
}
\description{
calculate the power of the test with J readers and n individuals with the condition when pilot data is not available
the accuracy is partial AUC
}
