% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.5.3.R
\name{MRMC_pilot}
\alias{MRMC_pilot}
\title{calculate the power of the test with J readers and N total patients when pilot data is available}
\usage{
MRMC_pilot(delta, J_pilot, N_pilot, J, N, MSTRP, MSTR, MSTP, alpha)
}
\arguments{
\item{delta}{the difference of accuracy of test 1 under the alternative hypothesis}

\item{J_pilot}{number of readers in the pilot study}

\item{N_pilot}{number of total patients in the pilot study}

\item{J}{number of readers in the current study}

\item{N}{number of total patients in the current study}

\item{MSTRP}{the modality-by- reader-by-patient interaction mean squares}

\item{MSTR}{the modality-by-reader interaction mean squares}

\item{MSTP}{the modality-by-patient interaction mean squares}

\item{alpha}{significance level}
}
\value{
the power of the test with J readers and N total patients when pilot data is available
}
\description{
calculate the power of the test with J readers and N total patients when pilot data is available
}
