% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{non_inferiority_PPV}
\alias{non_inferiority_PPV}
\title{Sample size for testing non-inferiority when the accuracy is PPV
using the method proposed by Moskowitz and Pepe(2006)}
\usage{
non_inferiority_PPV(PPV2, p, gamma, delta, alpha, beta)
}
\arguments{
\item{PPV2}{the conjectured value of PPV of test 2}

\item{p}{a vector with eight dimensions
For patients without the condition：
p[1] the proportion of patients testing positive on both tests
p[2] the proportion testing positive on test 1 and negative on test 2
p[3] the proportion testing positive on test 2 and negative on test 1
p[4] the proportion testing negative on both tests.
For patients with the condition:
p[5] the proportion of patients testing positive on both tests
p[6] the proportion testing positive on test 1 and negative on test 2
p[7] the proportion testing positive on test 2 and negative on test 1
p[8] the proportion testing negative on both tests}

\item{gamma}{a specific value of interest for rPPV under the alternative hypothesis}

\item{delta}{The null hypothesis is rPPV≤delta.}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}
}
\value{
n_one total sample size in a one-sided hypothesis testing
        n_two total sample sizein a two-sided hypothesis testing
}
\description{
Sample size for testing non-inferiority when the accuracy is PPV
using the method proposed by Moskowitz and Pepe(2006)
}
