% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{non_inferiority_rFPR}
\alias{non_inferiority_rFPR}
\title{Sample size for testing the non-inferiority when the accuracies are sensitivity and specificity
calculate the sample size needed for testing the rFPR parameter by constructing joint confidence intervals for rTPR and rFPR
proposed by Alonzo and Pepe(2002),for paired design}
\usage{
non_inferiority_rFPR(FPR2, FPPR, gamma, delta2, alpha, beta)
}
\arguments{
\item{FPR2}{the false positive rate for test 2}

\item{FPPR}{the probability that both tests 1 and 2 are positive for patients without the condition}

\item{gamma}{a specific value of interest for rFPR under the alternative hypothesis}

\item{delta2}{The null hypothesis is rFPR>delta2}

\item{alpha}{1-alpha is the level of the joint confidence interval for rTPR and rFPR}

\item{beta}{1-beta is the power of the test}
}
\value{
n_pro total sample size for prospective study
        n_retro total sample size for retrospective study
}
\description{
Sample size for testing the non-inferiority when the accuracies are sensitivity and specificity
calculate the sample size needed for testing the rFPR parameter by constructing joint confidence intervals for rTPR and rFPR
proposed by Alonzo and Pepe(2002),for paired design
}
