% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{non_inferiority_rTPR}
\alias{non_inferiority_rTPR}
\title{Sample size for testing the non-inferiority when the accuracies are sensitivity and specificity
calculate the sample size needed for testing the rTPR parameter by constructing joint confidence intervals for rTPR and rFPR
proposed by Alonzo and Pepe(2002),for paired design}
\usage{
non_inferiority_rTPR(TPR2, TPPR, gamma, delta1, alpha, beta)
}
\arguments{
\item{TPR2}{the sensitivity for test 2}

\item{TPPR}{the probability that both tests 1 and 2 are positive for patients with the condition}

\item{gamma}{a specific value of interest for rTPR under the alternative hypothesis}

\item{delta1}{The null hypothesis is rTPR<delta1}

\item{alpha}{1-alpha is the level of the joint confidence interval for rTPR and rFPR}

\item{beta}{1-beta is the power of the test}
}
\value{
n_pro total sample size for prospective study
        n_retro total sample size for retrospective study
}
\description{
Sample size for testing the non-inferiority when the accuracies are sensitivity and specificity
calculate the sample size needed for testing the rTPR parameter by constructing joint confidence intervals for rTPR and rFPR
proposed by Alonzo and Pepe(2002),for paired design
}
