% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3.R
\name{non_inferiority_Se}
\alias{non_inferiority_Se}
\title{Sample size for testing the non-inferiority when the accuracy is sensitivity}
\usage{
non_inferiority_Se(Se1, Se2, p, Delta, alpha, beta, R)
}
\arguments{
\item{Se1}{the conjectured value of sensitivity under the alternative hypothesis of test 1}

\item{Se2}{the conjectured value of sensitivity under the alternative hypothesis of test 2}

\item{p}{P(T1=1|T2=1)}

\item{Delta}{the smallest difference in accuracy which is not acceptable}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased subjects and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size(paired design)
}
\description{
Sample size for testing the non-inferiority when the accuracy is sensitivity
}
