\name{NSN-package}


\alias{NSN-package}
\alias{NSN}

\docType{package}

\title{
  NSN
}

\description{
  Fonctions de calcul de nombre de sujets nécessaires.
}

\details{
  \tabular{ll}{
    Package: \tab  NSN \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/11/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Autres packages developpes au CESPA :

  \code{\link{lazyMe}}, \code{\link{MDBprod}}, \code{\link{manageR}}, \code{\link{surveyscores}}, \code{\link{epiplus}}
}

\examples{

}
