% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.5.2.R
\name{sample_fixedreader_AUC_Blume}
\alias{sample_fixedreader_AUC_Blume}
\title{this function calculate the sample size needed for the fixed-reader MRMC design
using the method proposed by Blume(2009)}
\usage{
sample_fixedreader_AUC_Blume(theta, theta1, theta2, r, J, rho, alpha, beta, R)
}
\arguments{
\item{theta}{AUC of both tests under the null hypothesis}

\item{theta1}{AUC of test 1 under the alternative hypothesis}

\item{theta2}{AUC of test 2 under the alternative hypothesis}

\item{r}{the correlation between the tests because of the paired design}

\item{J}{number of readers}

\item{rho}{the correlation between different readers on the difference in ROC areas}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size needed
}
\description{
this function calculate the sample size needed for the fixed-reader MRMC design
using the method proposed by Blume(2009)
}
