% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.5.2.R
\name{sample_fixedreader_AUC_Obuchowski}
\alias{sample_fixedreader_AUC_Obuchowski}
\title{this function calculate the sample size needed for the fixed-reader MRMC design
using the method proposed by Obuchowski(1995)}
\usage{
sample_fixedreader_AUC_Obuchowski(
  a1_null,
  a2_null,
  a1_alter,
  a2_alter,
  rD,
  rN,
  J,
  rho,
  alpha,
  beta,
  R
)
}
\arguments{
\item{a1_null}{the binormal parameter a for test 1 under the null hypothesis}

\item{a2_null}{the binormal parameter a for test 2 under the null hypothesis}

\item{a1_alter}{the binormal parameter a for test 1 under the alternative hypothesis}

\item{a2_alter}{the binormal parameter a for test 2 under the alternative hypothesis}

\item{rD}{the correlation of the underlying bivariate binormal distribution for patients with the condition}

\item{rN}{the correlation of the underlying bivariate binormal distribution for patients without the condition}

\item{J}{number of readers}

\item{rho}{The difference between estimated correlation when the same patients are evaluated by different readers using the same test
and The estimated correlation when the same patients are evaluated by different readers using different tests}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n sample size of diseased subjects
        N total sample size needed
}
\description{
this function calculate the sample size needed for the fixed-reader MRMC design
using the method proposed by Obuchowski(1995)
}
