% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.2.4.R
\name{samplecompareAUC_any}
\alias{samplecompareAUC_any}
\title{Sample Size for Comparing Tests Area Under the ROC Curve
under any distribution}
\usage{
samplecompareAUC_any(theta, theta1, theta2, r, alpha, beta, R)
}
\arguments{
\item{theta}{AUC of both tests under the null hypothesis}

\item{theta1}{AUC of test 1 under the alternative hypothesis}

\item{theta2}{AUC of test 2 under the alternative hypothesis}

\item{r}{the correlation between the tests because of the paired design}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n_one sample size of diseased subjects in a one-sided hypothesis testing
        N_one total sample size in a one-sided hypothesis testing(paired design)
        n_two sample size of diseased subjects in a two-sided hypothesis testing
        N_two total sample size in a two-sided hypothesis testing(paired design)
}
\description{
Sample Size for Comparing Tests Area Under the ROC Curve
under any distribution
}
