% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.2.4.R
\name{samplecompareAUC_binormal}
\alias{samplecompareAUC_binormal}
\title{Sample Size for Comparing Tests Area Under the ROC Curve
under the binormal assumption}
\usage{
samplecompareAUC_binormal(
  a1_null,
  a2_null,
  a1_alter,
  a2_alter,
  rD,
  rN,
  alpha,
  beta,
  R
)
}
\arguments{
\item{a1_null}{the binormal parameter a for test 1 under the null hypothesis}

\item{a2_null}{the binormal parameter a for test 2 under the null hypothesis}

\item{a1_alter}{the binormal parameter a for test 1 under the alternative hypothesis}

\item{a2_alter}{the binormal parameter a for test 2 under the alternative hypothesis}

\item{rD}{the correlation of the underlying bivariate binormal distribution for patients with the condition}

\item{rN}{the correlation of the underlying bivariate binormal distribution for patients without the condition}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n_one sample size of diseased subjects in a one-sided hypothesis testing
        N_one total sample size in a one-sided hypothesis testing(paired design)
        n_two sample size of diseased subjects in a two-sided hypothesis testing
        N_two total sample size in a two-sided hypothesis testing(paired design)
}
\description{
Sample Size for Comparing Tests Area Under the ROC Curve
under the binormal assumption
}
