% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.2.7.R
\name{samplepartialAUC}
\alias{samplepartialAUC}
\title{Sample Size for Comparing Tests Partial Area Under the ROC Curve}
\usage{
samplepartialAUC(
  a1_null,
  b1_null,
  a2_null,
  b2_null,
  a1_alter,
  b1_alter,
  a2_alter,
  b2_alter,
  rD,
  rN,
  e1,
  e2,
  alpha,
  beta,
  R
)
}
\arguments{
\item{a1_null}{binormal parameter for test 1 under null hypothesis}

\item{b1_null}{binormal parameter for test 1 under null hypothesis}

\item{a2_null}{binormal parameter for test 2 under null hypothesis}

\item{b2_null}{binormal parameter for test 2 under null hypothesis}

\item{a1_alter}{binormal parameter for test 1 under alternative hypothesis}

\item{b1_alter}{binormal parameter for test 1 under alternative hypothesis}

\item{a2_alter}{binormal parameter for test 2 under alternative hypothesis}

\item{b2_alter}{binormal parameter for test 2 under alternative hypothesis}

\item{rD}{the correlation of the underlying bivariate binormal distribution for patients with the condition}

\item{rN}{the correlation of the underlying bivariate binormal distribution for patients without the condition}

\item{e1}{lower bound for partial area}

\item{e2}{upper bound for partial area}

\item{alpha}{significance level}

\item{beta}{1-beta is the power of the test}

\item{R}{the ratio of undiseased and diseased subjects}
}
\value{
n_one sample size of diseased subjects in a one-sided hypothesis testing
        N_one total sample size in a one-sided hypothesis testing(paired design)
        n_two sample size of diseased subjects in a two-sided hypothesis testing
        N_two total sample size in a two-sided hypothesis testing(paired design)
}
\description{
Sample Size for Comparing Tests Partial Area Under the ROC Curve
}
