% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.1.1.R
\name{sampleSeSp}
\alias{sampleSeSp}
\title{Sample Size Calculations for Estimating Sensitivity and Specificity}
\usage{
sampleSeSp(Se, Sp, alpha, beta, L, p)
}
\arguments{
\item{Se}{the conjectured sensitivity of the test}

\item{Sp}{the conjectured specificity of the test}

\item{alpha}{1-alpha is the size of the desired confidence interval}

\item{beta}{1-beta is the desired power}

\item{L}{the desired width of one-half of the CI}

\item{p}{the prevalence of the condition in the population}
}
\value{
n.normal is a more precise estimation of the number of patients with condition
        n.rough is a rough estimation of the number of patients with condition
        N.normal is a more precise estimation of the total number of patients
        N.rough is a rough estimation of the total number of patients
}
\description{
Sample Size Calculations for Estimating Sensitivity and Specificity
}
