\name{samplesize-package}
\alias{samplesize-package}
\docType{package}
\title{
 Computes sample size for several two-sample tests
}
\description{
 Computes sample size for independent and paired Student's t-test,
 Student's t-test with Welch-approximation, Wilcoxon-Mann-Whitney test
 with and without ties on ordinal data
}
\details{
\tabular{ll}{
Package: \tab samplesize\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2010-03-19\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
 n.paired.t.test(): sample size for paired Student's t-test

 n.indep.t.test.eq(): sample size for indepent Student's t-test with equal group size

 n.indep.t.test.neq(): sample size for indepent Student's t-test with different group size

 n.welch.test(): sample size for Student's t.test with Welch approximation
 
 n.wilcox.ord(): sample size for Wilcoxon-Mann-Whitney test with and
 without ties
}
\author{
Ralph Scherer

Maintainer: Ralph Scherer <scherer.ralph@mh-hannover.de>
}
\references{
 Bock J., Bestimmung des Stichprobenumfangs fuer biologische Experimente und kontrollierte klinische Studien. Oldenbourg 1998

 Zhao YD, Rahardja D, Qu Yongming. Sample size calculation for the Wilcoxon-Mann-Whitney test adjusting for ties. Statistics in Medicine 2008; 27:462-468
}
\keyword{htest}

