\name{sampleSize-package}


\alias{sampleSize-package}
\alias{sampleSize}

\docType{package}

\title{
  sampleSize
}

\description{
  Sample size calculations.
}

\details{
  \tabular{ll}{
    Package: \tab  sampleSize \cr
    Type: \tab
    Package\cr
    Version: \tab  1.0 \cr
    Date: \tab  01/11/2022 \cr
    License: \tab void \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages by the same author :

  \code{\link{lazyMe}}, \code{\link{dataDesc}}, \code{\link{manageR}}, \code{\link{surveyscores}}, \code{\link{epiplus}}
}

\examples{

}
