\name{samplesize-package}
\alias{samplesize-package}
\docType{package}
\title{
 Computes sample size for several two-sample tests
}
\description{
 Computes sample size for independent and paired Student's t-test,
 Student's t-test with Welch-approximation, Wilcoxon-Mann-Whitney test
 with and without ties on ordinal data
}
\details{
\tabular{ll}{
Package: \tab samplesize\cr
Type: \tab Package\cr
Version: \tab 0.2-4\cr
Date: \tab 2016-12-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
 n.ttest(): sample size for Student's t-test and t-test with Welch approximation
 
 n.wilcox.ord(): sample size for Wilcoxon-Mann-Whitney test with and
 without ties
}
\author{
Ralph Scherer

Maintainer: Ralph Scherer <shearer.ra@gmail.com>
}
\references{
 Bock J., Bestimmung des Stichprobenumfangs fuer biologische Experimente und kontrollierte klinische Studien. Oldenbourg 1998

 Zhao YD, Rahardja D, Qu Yongming. Sample size calculation for the Wilcoxon-Mann-Whitney test adjusting for ties. Statistics in Medicine 2008; 27:462-468
}
\keyword{htest}

